Move product related methods into products service
