+++ src/services/Api.php
@@ -401,79 +401,6 @@ class Api extends Component
     }
 
     /**
-     * @param array|Product[] $products
-     * @return array
-     * @since 6.0.0
-     */
-    public function eagerLoadMetafieldsForProducts(array $products): array
-    {
-        return $this->_eagerLoadTypeOnProducts($products, 'Metafield', function($product, $metafields) {
-            if (!empty($metafields)) {
-                // Squash to key/value array
-                $metafields = array_combine(ArrayHelper::getColumn($metafields, 'key'), ArrayHelper::getColumn($metafields, 'value'));
-            }
-            $product->setMetafields($metafields);
-        });
-    }
-
-    /**
-     * @param array|Product[] $products
-     * @return array
-     * @since 6.0.0
-     */
-    public function eagerLoadImagesForProducts(array $products): array
-    {
-        return $this->_eagerLoadTypeOnProducts($products, 'MediaImage', function($product, $images) {
-            $product->setImages($images);
-        });
-    }
-
-    /**
-     * @param array $products
-     * @return array
-     * @since 6.0.0
-     */
-    public function eagerLoadVariantsForProducts(array $products): array
-    {
-        return $this->_eagerLoadTypeOnProducts($products, 'ProductVariant', function($product, $variants) {
-            $product->setVariants($variants);
-        });
-    }
-
-    /**
-     * @param array $products
-     * @param string $type
-     * @param callable $callback
-     * @return array
-     */
-    private function _eagerLoadTypeOnProducts(array $products, string $type, callable $callback): array
-    {
-        $productIds = ArrayHelper::getColumn($products, 'shopifyGid');
-        $data = $this->getShopifyDataByType($type, $productIds);
-
-        if (empty($data)) {
-            foreach ($products as $product) {
-                $callback($product, []);
-            }
-        }
-
-        // Group images by product ID
-        $data = collect($data)->groupBy('__parentId');
-
-        foreach ($products as $product) {
-            $productData = $data->get($product->shopifyGid, []);
-            if (empty($productData)) {
-                $callback($product, []);
-                continue;
-            }
-
-            $callback($product, $productData->all());
-        }
-
-        return $products;
-    }
-
-    /**
      * Returns or sets up a Rest API client.
      *
      * @return Graphql
